added new filters

+++ kg1/kg1/Filters.cs
@@ -29,7 +29,7 @@ namespace kg1
 
         public int clamp(int value, int min, int max)
         {
-            if (value < min) return min;
+            if (value < min) return (byte)min;
             if (value > max) return max;
             return value;
         }
